backend/pyproject.toml
[project]
name = "fallout-shelter-fast-api"
version = "2.10.8"
authors = [{ name = "ElderEvil", email = "elder.evil.dev@proton.me" }, ]
description = "Fallout Shelter FastAPI Game is a web-based simulation game where the player manages a vault full of dwellers, balancing their needs and resources to keep the vault thriving."
requires-python = ">=3.12,<3.14"
dependencies = [
    "alembic>=1.18.1,<2.0",
    "asyncpg>=0.30.0,<1.0",
    "celery>=5.6.1,<6.0",
    "sqlalchemy-celery-beat>=0.7.4",
    "faker>=40.1.2",
    "fastapi[standard]>=0.128.0",
    "minio>=7.2.20,<8.0",
    "boto3>=1.34.0,<2.0",
    "botocore>=1.34.0,<2.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0,<13",
    "psycopg>=3.3.2,<4",
    "pydantic>=2.11.10,<2.13",
    "pydantic-settings>=2.12.0,<3",
    "pydantic-extra-types>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0,<8",
    "sqladmin[full]>=0.22.0",
    "sqlmodel>=0.0.31,<1",
    "httpcore>=1.0.9",
    "httptools>=0.7.1,<1",
    "greenlet>=3.3.1",
    "pydantic-ai-slim[openai, anthropic, logfire]>=1.39.0",
    "aiosmtplib>=5.1.0",
    "python-json-logger>=4.0.0",
    "psycopg2>=2.9.11",
    "uvloop>=0.21.0; sys_platform != 'win32'",
    "fastapi-guard>=4.2.2,<5",
]

[dependency-groups]
test = [
    "aiosqlite>=0.22.1",
    "coverage>=7.13.2",
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "celery-types>=0.24.0",
]
lint = [
    "prek>=0.3.0",
    "ruff>=0.14.14",
    "ty>=0.0.13",
]
perf = [
    "locust>=2.43.1",
]

[tool.uv]
default-groups = ["test", "lint", "perf"]

[build-system]
requires = ["setuptools >= 75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["py.typed"]

[tool.ruff]
line-length = 120
target-version = "py313"
exclude = [".git", "__pycache__", ".pytest_cache", "alembic", "app/cli"]
lint.select = [
    "FAST",
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "EM",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "INP",
    "PIE",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SIM",
    "TID",
    "INT",
    "ARG",
    "PTH",
    "PGH",
    "PL",
    "TRY",
    "RUF",
]

lint.mccabe.max-complexity = 12

lint.ignore = [
    "COM812",
    "ISC001",
    "B008",
    "PTH123",
    "S311",
    "PLR2004",
    "PLC0415",
    "PT023",
    "RUF012",
    "N805",
    "E712",
    "UP037",
    "UP045",
    "UP046",
    "TRY301",
    "TRY300",
    "G004",
    "FBT001",  # Boolean-typed positional argument
    "FBT002",  # Boolean default positional argument
    "EM101",   # Exception must not use string literal
    "EM102",   # Exception must not use f-string literal
    "TRY003",  # Avoid specifying long messages outside exception class
    "PLR0913", # Too many arguments in function definition
]

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"app/tests/**" = ["ARG001", "F401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
pythonpath = ["."]
addopts = [
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=json",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/alembic/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2
